api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
email-validator
python-dotenv
python-multipart
bcrypt
